requirements.txt
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
docutils==0.21.2
et_xmlfile==2.0.0
fonttools==4.59.0
greenlet==3.2.3
h11==0.16.0
id==1.5.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
keyring==25.6.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.10.3
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.2.22
numpy==2.3.1
openpyxl==3.2.0b1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pycparser==2.22
Pygments==2.19.2
pyloco==0.0.139
PyMySQL==1.1.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
selenium==4.34.2
# Editable install with no version control (semoosa==1.0.0)
-e d:\python3\account
setuptools==80.9.0
SimpleWebSocketServer==0.1.2
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.42
trio==0.30.0
trio-websocket==0.12.2
twine==6.1.0
typing==3.7.4.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
ushlex==0.99.1
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
xlrd==2.0.2
xlsxwriter==3.2.9
